+++ ui/icon/src/icon.ts
@@ -1,12 +1,16 @@
-import {AlwatrElement} from '@alwatr/element';
+import {
+  AlwatrDummyElement,
+  unsafeSVG,
+  customElement,
+  property,
+  state,
+  html,
+  css,
+  DirectionMixin,
+} from '@alwatr/element';
 import {fetch} from '@alwatr/fetch';
-import {html, css} from 'lit';
-import {customElement} from 'lit/decorators.js';
-import {property} from 'lit/decorators/property.js';
-import {state} from 'lit/decorators/state.js';
-import {unsafeSVG} from 'lit/directives/unsafe-svg.js';
 
-import type {PropertyValues, HTMLTemplateResult} from 'lit';
+import type {PropertyValues, HTMLTemplateResult} from '@alwatr/element';
 
 declare global {
   interface HTMLElementTagNameMap {
@@ -20,7 +24,7 @@ declare global {
  * @attr {boolean} flip-rtl
  */
 @customElement('alwatr-icon')
-export class AlwatrIcon extends AlwatrElement {
+export class AlwatrIcon extends DirectionMixin(AlwatrDummyElement) {
   static override styles = css`
     :host {
       display: inline-block;
@@ -65,7 +69,7 @@ export class AlwatrIcon extends AlwatrElement {
   protected _icon?: HTMLTemplateResult;
 
   override render(): unknown {
-    this._logger.logMethod('render');
+    super.render();
     return this._icon;
   }
 
@@ -73,7 +77,7 @@ export class AlwatrIcon extends AlwatrElement {
     if (changedProperties.has('name') || changedProperties.has('urlPrefix')) {
       this._fetchIcon();
     }
-    return changedProperties.has('_icon') && this._icon != null;
+    return super.shouldUpdate(changedProperties) && changedProperties.has('_icon') && this._icon != null;
   }
 
   protected async _fetchIcon(): Promise<void> {
@@ -98,8 +102,11 @@ export async function preloadIcon(
   const url = urlPrefix + name + '.svg';
   const response = await fetch({
     url,
+    timeout: 6000,
+    retry: 5,
     removeDuplicate: 'auto',
     cacheStrategy: 'cache_first',
+    cache: 'force-cache',
   });
 
   if (response.ok !== true) {