+++ ui/ui-kit/CHANGELOG.md
@@ -3,6 +3,53 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.27.0](https://github.com/AliMD/alwatr/compare/v0.26.0...v0.27.0) (2022-12-29)
+
+### Bug Fixes
+
+- **chat-avatar:** grow ([d935030](https://github.com/AliMD/alwatr/commit/d9350303d9345729ab83f410182afeb1b7823294))
+- **chat-avatar:** size with new root fs ([63c8f36](https://github.com/AliMD/alwatr/commit/63c8f3617856a2ee65ccf7b398d5635349afb019))
+- **chat-footer:** footer padding, demo \_changeLocale ([6454599](https://github.com/AliMD/alwatr/commit/6454599272bc0c1b01c223274f48b8de99655130))
+- **standard-icon-button:** style ([0cfe9ee](https://github.com/AliMD/alwatr/commit/0cfe9ee6ea3e782b95d8953686a25fdc6c7d2847))
+- **ui-kit/chat-footer:** styles ([38e3bbc](https://github.com/AliMD/alwatr/commit/38e3bbc5d35242466cd4e33bf4bcacd5c6a87368))
+- **ui-kit/chat-text-input:** direction mixin ([8f00eac](https://github.com/AliMD/alwatr/commit/8f00eac04fe64fe5aa933a37fddce93c24c1bc58))
+- **ui-kit/icon-button:** flip-rtl attr ([9607493](https://github.com/AliMD/alwatr/commit/9607493754ad454c27f26d0eba883ca4488fca67))
+- **ui-kit:** font sizes ([2569644](https://github.com/AliMD/alwatr/commit/25696442da35ba570682c31157489d5802153fbf))
+- **ui:** root sizes ([2bdf2d5](https://github.com/AliMD/alwatr/commit/2bdf2d5712905d75491f61a5d94322395987b6ab))
+
+### Features
+
+- **chat-\*:** scrollbar and gaps ([c91e2d6](https://github.com/AliMD/alwatr/commit/c91e2d6ec5a4dcc80b2ae71f800dae46cdd8c1f8))
+- **chat-avatar:** new style ([ea11421](https://github.com/AliMD/alwatr/commit/ea114213ad811206897c57f885b2973182a62f94))
+- **chat-bubble:** rewrite with new styles ([f4c5f01](https://github.com/AliMD/alwatr/commit/f4c5f0154935a1df77493f309ac9ce9b072ce5a2))
+- **chat-bubble:** side attrib ([5c36f8d](https://github.com/AliMD/alwatr/commit/5c36f8d5093659b7d6b6a251822294530f235fd0))
+- **chat-list:** new element for list of chats ([3ddae05](https://github.com/AliMD/alwatr/commit/3ddae05cd0bcebb855327f0f449d3ee7c7e576cd))
+- **chat-list:** scrollbar style ([a02074d](https://github.com/AliMD/alwatr/commit/a02074d894d3d812bfe8fb63be37f0edde931c00))
+- **chat-list:** self property and gap and spaces ([9d9edca](https://github.com/AliMD/alwatr/commit/9d9edca5a9a1c4ae1cf9f085ca3da92dfad7acae))
+- **chat-message:** ChatTextMessage type ([227dda3](https://github.com/AliMD/alwatr/commit/227dda3611049f787bfaee2f4443420b005475e2))
+- **chat-message:** gap and spaces ([0b1043a](https://github.com/AliMD/alwatr/commit/0b1043a9a9ba2c67b664ca2edc7ad8610b05c210))
+- **chat-message:** new chat element ([12f9b72](https://github.com/AliMD/alwatr/commit/12f9b724ec35a55d5e7a1d8495dd11ba0f6a77a4))
+- **chat-text-input:** inputElement ([4abfdd6](https://github.com/AliMD/alwatr/commit/4abfdd6c4e73e5db588e6f099fc9d5e1b8cd06db))
+- **chat-text-input:** style ([8c302cc](https://github.com/AliMD/alwatr/commit/8c302ccd7045a91f596a6ddae601ae87a26a8fc5))
+- **chat:** use new color system ([045b188](https://github.com/AliMD/alwatr/commit/045b18891ca8b9ba0a556be6c621747cb8d0a639))
+- **chat:** use new spacing system ([4e0d44e](https://github.com/AliMD/alwatr/commit/4e0d44e419973e28d5584e609f8efbc68f07fcd2))
+- **chat:** use system shape var ([dcff276](https://github.com/AliMD/alwatr/commit/dcff276ec28e345306f1a278a69bd64c05c5e130))
+- **icon-button:** enhance styles and motion ([4b4a301](https://github.com/AliMD/alwatr/commit/4b4a301f53a059cc5faa4a75b22fc079f750bc38))
+- new ui design system ([8c9c8ba](https://github.com/AliMD/alwatr/commit/8c9c8ba29c477838f3e7034b902f4e0dafabd17c))
+- **page-chat:** new demo ([451d63f](https://github.com/AliMD/alwatr/commit/451d63fc0b904bb780c898c2909b1d7a116b353f))
+- **ui-kit/chat-footer:** build base ([30cf89a](https://github.com/AliMD/alwatr/commit/30cf89aac3a68e0fc028eb7a8a227efce85ba892))
+- **ui-kit/chat-footer:** component ([7041047](https://github.com/AliMD/alwatr/commit/70410476ff7846b14873dfdf6299d528e1de2f61))
+- **ui-kit/chat-footer:** new component ([5151d1c](https://github.com/AliMD/alwatr/commit/5151d1c50c1fd63e1bee479220a4363a9ab800e3))
+- **ui-kit/chat-text-input:** flex grow ([b8837ff](https://github.com/AliMD/alwatr/commit/b8837ff9772959a585e4862f48b9b5da390da561))
+- **ui-kit/chat-text-input:** improve ([28f08a9](https://github.com/AliMD/alwatr/commit/28f08a91ff49ac4567c86fea7d3bd4458d69cc4a))
+- **ui-kit/chat-text-input:** translations ([3d03d3d](https://github.com/AliMD/alwatr/commit/3d03d3d56d9fef34b35c76568fb9d7ec7069f1ee))
+- **ui-kit/chat-text-input:** typography ([e67113e](https://github.com/AliMD/alwatr/commit/e67113ec2e3c8e32a5e700162755995d852aec66))
+- **ui-kit/icon-button:** component ([7db86bf](https://github.com/AliMD/alwatr/commit/7db86bfa678d64237951a8c8002c38dd51d3d082))
+- **ui-kit/icon-button:** icon flip-rtl ([f6427b6](https://github.com/AliMD/alwatr/commit/f6427b62315389cea308f2a522e358af8b9838cd))
+- **ui-kit:** elevation sample ([ccd9e8a](https://github.com/AliMD/alwatr/commit/ccd9e8a9c7cc94a830a03bf11ddbf46891300f68))
+- **ui/chat:** new alwatr-chat element ([f711e12](https://github.com/AliMD/alwatr/commit/f711e12317a263d6e1d4bfc52dae54348cf64923))
+- **ui/ui-kit:** chat text input component ([586aefd](https://github.com/AliMD/alwatr/commit/586aefd7fcbfd520f36f035415dbb76d5ffa324f))
+
 # [0.26.0](https://github.com/AliMD/alwatr/compare/v0.25.0...v0.26.0) (2022-12-22)
 
 ### Bug Fixes